requirements.txt
python-telegram-bot==20.7
groq==0.11.0
python-dotenv==1.0.0
aiohttp==3.9.1
schedule==1.2.1
